requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Statistical and time series
statsmodels>=0.14.0
arch>=6.0.0

# Hidden Markov Models
hmmlearn>=0.3.0

# Alternative HMM library (optional, more features)
# pomegranate>=0.14.0

# Data acceleration
numba>=0.57.0

# Exchange APIs and crypto data
ccxt>=4.0.0
python-binance>=1.0.17

# Configuration and validation
pyyaml>=6.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# CLI
typer>=0.9.0
rich>=13.0.0
click>=8.1.0

# Time handling
pytz>=2023.3
python-dateutil>=2.8.0

# Technical indicators
TA-Lib>=0.4.25

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# ML tracking (optional)
mlflow>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Data formats
pyarrow>=12.0.0  # for parquet
openpyxl>=3.1.0  # for excel

# HTTP requests (for APIs)
requests>=2.31.0
aiohttp>=3.8.0

# Utilities
tqdm>=4.65.0
pathlib2>=2.3.7; python_version < '3.4'

# Optional: Advanced features
# lightgbm>=4.0.0
# xgboost>=1.7.0
# optuna>=3.3.0

# Docker support
gunicorn>=21.0.0

# Security
cryptography>=41.0.0

# Memory optimization
psutil>=5.9.0

# CLI & Console
ccxt>=4.3.0
rich>=13.7.0
typer>=0.12.0
